requirements.txt
asgiref==3.8.1
certifi==2024.6.2
Django==5.0.6
django-elasticsearch-dsl==8.0
djangorestframework==3.15.2
elastic-transport==8.13.1
elasticsearch==8.14.0
elasticsearch-dsl==8.14.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
